fix overflow issue

+++ 0063-Unique-Paths-II/cpp-0063/main.cpp
@@ -22,7 +22,7 @@ public:
         if(!n || obstacleGrid[0][0])
             return 0;
 
-        vector<vector<int>> dp(m, vector<int>(n, -1));
+        vector<vector<long long>> dp(m, vector<int>(n, -1));
         dp[0][0] = 1;
         for(int j = 1; j < n; j ++)
             if(obstacleGrid[0][j])
